Flytter typer over til egen types.ts slik gjort med andre typer
