+++ packages/server/src/views/scripts.ts
@@ -1,10 +1,10 @@
 import html, { json, Template, unsafeHtml } from "decorator-shared/html";
 import { buildHtmlElementString } from "../lib/html-element-string-builder";
-import { HtmlElementProps } from "decorator-shared/types";
+import { DecoratorDataProps, HtmlElementProps } from "decorator-shared/types";
 import { env } from "../env/server";
 import type { Manifest as ViteManifest } from "vite";
 import { buildCdnUrl } from "../urls";
-import { buildDecoratorData, DecoratorDataProps } from "../decorator-data";
+import { buildDecoratorData } from "../decorator-data";
 
 const ENTRY_POINT_PATH = "src/main.ts";
